// src/app/core/interfaces/clients/iuser-client.ts
export interface IUserClient {
    ID_USUARIO_CLIENTE: string;  
    ID_CLIENTE: string;           
    NOMBRE: string;
    APELLIDO: string;
    CARGO: string;
    USUARIO_DOMINIO: string;
    CORREO: string;
    CC: string;                   
    TELEFONO: string;
    PASSWORD: string;     
    EMPRESA:string;      
}